requirements.txt
opencv-python
Pillow
pip
numpy
setuptools
requests